requirements.txt
python-telegram-bot==20.3
python-dotenv
psycopg2-binary